Consolidating test files
