+++ tests/bundle/test_bundle_download.py
@@ -34,6 +34,7 @@ from tests.test_utils import (
     skip_if_downloading_fails,
     skip_if_no_cuda,
     skip_if_quick,
+    skip_if_windows
 )
 
 _, has_huggingface_hub = optional_import("huggingface_hub")
@@ -94,6 +95,47 @@ TEST_CASE_10 = [
     "https://github.com/Project-MONAI/MONAI-extra-test-data/releases/download/0.8.1/test_bundle_v0.1.3.zip",
     {"model.pt": "27952767e2e154e3b0ee65defc5aed38", "model.ts": "97746870fe591f69ac09827175b00675"},
 ]
+
+TEST_CASE_NGC_1 = [
+    "spleen_ct_segmentation",
+    "0.3.7",
+    None,
+    "monai_spleen_ct_segmentation",
+    "models/model.pt",
+    "b418a2dc8672ce2fd98dc255036e7a3d",
+]
+TEST_CASE_NGC_2 = [
+    "monai_spleen_ct_segmentation",
+    "0.3.7",
+    "monai_",
+    "spleen_ct_segmentation",
+    "models/model.pt",
+    "b418a2dc8672ce2fd98dc255036e7a3d",
+]
+
+TESTCASE_NGC_WEIGHTS = {
+    "key": "model.0.conv.unit0.adn.N.bias",
+    "value": torch.tensor(
+        [
+            -0.0705,
+            -0.0937,
+            -0.0422,
+            -0.2068,
+            0.1023,
+            -0.2007,
+            -0.0883,
+            0.0018,
+            -0.1719,
+            0.0116,
+            0.0285,
+            -0.0044,
+            0.1223,
+            -0.1287,
+            -0.1858,
+            0.0460,
+        ]
+    ),
+}
 
 
 class TestDownload(unittest.TestCase):
@@ -417,6 +459,32 @@ class TestDownloadLargefiles(unittest.TestCase):
                 for file in ["model.pt", "model.ts"]:
                     file_path = os.path.join(tempdir, bundle_name, f"models/{file}")
                     self.assertTrue(check_hash(filepath=file_path, val=hash_val[file]))
+
+
+@skip_if_windows
+class TestNgcBundleDownload(unittest.TestCase):
+    @parameterized.expand([TEST_CASE_NGC_1, TEST_CASE_NGC_2])
+    @skip_if_quick
+    def test_ngc_download_bundle(self, bundle_name, version, remove_prefix, download_name, file_path, hash_val):
+        with skip_if_downloading_fails():
+            with tempfile.TemporaryDirectory() as tempdir:
+                download(
+                    name=bundle_name, source="ngc", version=version, bundle_dir=tempdir, remove_prefix=remove_prefix
+                )
+                full_file_path = os.path.join(tempdir, download_name, file_path)
+                self.assertTrue(os.path.exists(full_file_path))
+                self.assertTrue(check_hash(filepath=full_file_path, val=hash_val))
+
+                model = load(
+                    name=bundle_name, source="ngc", version=version, bundle_dir=tempdir, remove_prefix=remove_prefix
+                )
+                assert_allclose(
+                    model.state_dict()[TESTCASE_NGC_WEIGHTS["key"]],
+                    TESTCASE_NGC_WEIGHTS["value"],
+                    atol=1e-4,
+                    rtol=1e-4,
+                    type_test=False,
+                )
 
 
 if __name__ == "__main__":
